requirements.txt
transformers>=4.35.0
pandas>=2.0.0
datasets>=2.14.0
peft>=0.6.0
trl>=0.7.4
torch>=2.0.0
lm-eval>=0.3.0
pyyaml>=6.0
bitsandbytes>=0.41.0
oauthlib>=3.2.0
filelock>=3.12.0
accelerate>=0.23.0
wandb>=0.15.0
langdetect
immutabledict
git+https://github.com/felipemaiapolo/tinyBenchmarks
requests>=2.28.0
openai>=1.1.0